Adds API call to persist item 'done'
Begins action in the `listr-list/item` component, which sends the
action to the `listr-list` component, which sends the action up to the
`list` route, which toggles item.done, and fires a PUT to the api.

In addition, we've modified the listItemCompleted attribute (in the
`listr-list/item` component) to be an Ember.computed attribute so
the css is tied directly to the value of the item.done value (instead
of an arbitrary value that just depends on the number of clicks).