Adds list model and updates model hook
Previously we just had dummy data in the model hook that was hard
coded. This change updates the model hook to pull data from our
listr-api instead. In order to do that we needed to tell ember what
to expect by creating a model for `list` with the correct attributes.
List `items` stopped showing up, so I'm about to check the api to make
sure we're serializing them out and that Ember is expecting them.